Separated out new texture-related structures and texture data generation from texture copying tests into a new util file to make them accessible to all gfx unit tests
